requirements.txt
aiofiles==0.4.0
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
certifi==2017.7.27.1
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
coverage==5.3
cryptography==3.1.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-crispy-forms==1.9.2
django-google-maps==0.12.2
django-multiselectfield==0.1.12
django-phonenumber-field==5.0.0
django-rest-framework==0.1.0
django-sendgrid-v5==0.9.0
django-tellme==0.6.5
djangorestframework==3.12.1
factory-boy==3.1.0
Faker==4.14.0
future==0.18.2
h11==0.11.0
httpcore==0.12.0
httpx==0.16.1
idna==2.7
iniconfig==1.1.1
isort==5.6.4
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.4
Paperboy==1.0.1
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-mock==3.3.1
python-dateutil==2.8.1
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2017.2
requests==2.20.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
selenium==3.141.0
sendgrid==6.4.7
simplejson==3.17.2
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
starlette==0.12.13
text-unidecode==1.3
toml==0.10.1
urllib3==1.24.3
wrapt==1.12.1
